refactor : 영화 인기순 조회 API 클린 아키텍처 도입 - 전체 패키지 구조를 data, domain, presentaion 레이어와 adapters 패키지로 구분 - 영화 인기순 조회 API에 클린 아키텍처 도입 - TODO    1. 파이어베이스를 호출하는 유저 정보 조회 및 로그인 구조 리팩터링    2. OAuth 로그인에 Strategy 패턴 도입    3. Dependency Injection 도입
