+++ src/data/DataResult.ts
