+++ src/data/datasource/MovieDataSource.ts
@@ -1,0 +1,8 @@
+import { MovieResponse } from "../entities/MovieEntity";
+import DataResult from "../DataResult";
+
+export interface MovieDataSource {
+  getMovieResponseOrderedByPopular(
+    pageId: number
+  ): Promise<DataResult<MovieResponse>>;
+}
