+++ src/domain/repository/MovieRepository.ts
@@ -1,0 +1,8 @@
+import { MovieGameInfo } from "../model/movie/MovieGameInfo";
+
+export interface MovieRepository {
+  getNextPopularMovies(
+    page: number,
+    baseIndex: number
+  ): Promise<MovieGameInfo[]>;
+}
